requirements.txt
# Core libraries
telethon>=1.34.0
sqlalchemy>=2.0.30
schedule>=1.2.1
python-dotenv~=1.1.1

# Database driver
# SQLite by default (built into Python).
# For PostgreSQL:
# psycopg2-binary>=2.9.9
# For MySQL/MariaDB:
# pymysql>=1.1.1
